refactor jira models
