+++ pkg/jira/models/BUILD.bazel
@@ -3,7 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "models",
     srcs = ["models.go"],
-    importpath = "github.com/stebennett/squad-dashboard/cmd/jiracollector/models",
+    importpath = "github.com/stebennett/squad-dashboard/pkg/jira/models",
     visibility = ["//visibility:public"],
-    deps = ["//pkg/jira/models"],
 )
