--- pkg/jiramodels/BUILD.bazel
@@ -1,8 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "jiramodels",
-    srcs = ["jiramodels.go"],
-    importpath = "github.com/stebennett/squad-dashboard/pkg/jiramodels",
-    visibility = ["//visibility:public"],
-)
