+++ pkg/jirarepository/postgresjirarepository.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/lib/pq"
 	"github.com/stebennett/squad-dashboard/pkg/calculatormodels"
-	"github.com/stebennett/squad-dashboard/pkg/jiramodels"
+	jiramodels "github.com/stebennett/squad-dashboard/pkg/jira/models"
 )
 
 type PostgresJiraRepository struct {
